Move folder from gl to canvas
the folder contained utility used for canvas 2d not WebGL. Renamed it not to confuse readers
