+++ web/src/app/timeline/components/canvas/rendering-loop-manager.ts
